packages/python/pyproject.toml
[project]
name = "lumina-ai-python"
version = "1.0.0"
description = "Lumina AI Python Services - AI/ML operations and data processing"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "langchain>=0.0.350",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.2",
    "numpy>=1.24.3",
    "pandas>=2.1.3",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "httpx>=0.25.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

